% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k2.R
\name{k2}
\alias{k2}
\title{k2 algorithm for structure learning}
\usage{
k2(variables, x, u, show = FALSE)
}
\arguments{
\item{variables}{A character vector specifying the variable names in order.}

\item{x}{data.frame containing the variables.}

\item{u}{The maximum number of parents for each variable.}

\item{show}{Logical, indicating whether to display intermediate steps (default is FALSE).}
}
\value{
A Bayesian network structure represented as a graph.
}
\description{
k2 algorithm for structure learning
}
\details{
K2 Algorithm is a structure learning algorithm for bayesian network, which is usefull
when the orders of the nodes are availeable,
}
\note{
the algorithm uses the logarithm of the score function.
}
\examples{
require(bnlearn)
require(Rgraphviz)

x1 <- c(1,1,0,1,0,0,1,0,1,0)
x2 <- c(0,1,0,1,0,1,1,0,1,0)
x3 <- c(0,1,1,1,0,1,1,0,1,0)
df <- data.frame(x1, x2, x3)

arti_k2 <- k2(c("x1", "x2", "x3"), df, 3, TRUE)
graphviz.plot(arti_k2, shape = "ellipse")

}
\references{
Cooper, G. F., & Herskovits, E. (1992). A Bayesian method for the induction of probabilistic networks
from data. Machine learning, 9(4), 309-347.
}
\seealso{
\code{\link{graphviz.plot}}, \code{\link{empty.graph}}, \code{\link{arcs}}
}
\keyword{algorithm}
\keyword{bayesian}
\keyword{k2}
\keyword{learning}
\keyword{network}
\keyword{structure}
