% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_fitted_plot.R
\name{pie_fitted_plot}
\alias{pie_fitted_plot}
\title{Pie Chart with Fitted Curve for Multiple Parents}
\usage{
pie_fitted_plot(fitted_dag, node)
}
\arguments{
\item{fitted_dag}{`bn.fit` object}

\item{node}{name of the node to plot the probabilities}
}
\value{
`ggplot2` plot
}
\description{
Pie Chart with Fitted Curve for Multiple Parents
}
\details{
The `fitted_dag` argument should be a bn.fit object obtained by estimating
parameters using methods such as MLE or Bayesian methods. The `node` argument
specifies the name of the node for which the pie chart and fitted curve will 
be plotted. The function is particularly useful when there are multiple 
parents in the k2 algorithm.
}
\examples{
require(bnlearn)
require(ggplot2)
data(insurance)
modelstring <-  
 paste0("[Age][Mileage][SocioEcon|Age]",
   "[GoodStudent|Age:SocioEcon]",
   "[RiskAversion|Age:SocioEcon][OtherCar|SocioEcon]",
   "[VehicleYear|SocioEcon:RiskAversion]",
   "[MakeModel|SocioEcon:RiskAversion]",
   "[SeniorTrain|Age:RiskAversion]",
   "[HomeBase|SocioEcon:RiskAversion]",
   "[AntiTheft|SocioEcon:RiskAversion]",
   "[RuggedAuto|VehicleYear:MakeModel]",
   "[Antilock|VehicleYear:MakeModel]",
   "[DrivingSkill|Age:SeniorTrain]",
   "[CarValue|VehicleYear:MakeModel:Mileage]",
   "[Airbag|VehicleYear:MakeModel]",
   "[DrivQuality|RiskAversion:DrivingSkill]",
   "[Theft|CarValue:HomeBase:AntiTheft]",
   "[Cushioning|RuggedAuto:Airbag]",
   "[DrivHist|RiskAversion:DrivingSkill]",
   "[Accident|DrivQuality:Mileage:Antilock]",
   "[ThisCarDam|RuggedAuto:Accident]",
   "[OtherCarCost|RuggedAuto:Accident]",
   "[MedCost|Age:Accident:Cushioning]",
   "[ILiCost|Accident]",
   "[ThisCarCost|ThisCarDam:Theft:CarValue]",
   "[PropCost|ThisCarCost:OtherCarCost]")
dag = model2network(modelstring)
ins_fit <- bn.fit(dag,insurance)
graphviz.plot(dag)

pie_fitted_plot(ins_fit,"GoodStudent")
}
