% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_fitted_plot.R
\name{pie_fitted_plot}
\alias{pie_fitted_plot}
\title{Title}
\usage{
pie_fitted_plot(fitted_dag, node)
}
\arguments{
\item{fitted_dag}{bn.fit object}

\item{node}{name of the node to plot the probabilities}
}
\value{
ggplot2 plot
}
\description{
Title
}
\details{
Additional details...
}
\examples{
require(bnlearn)
required(ggplot2)
data(insurance)
modelstring <-  
 paste0("[Age][Mileage][SocioEcon|Age]",
   "[GoodStudent|Age:SocioEcon]",
   "[RiskAversion|Age:SocioEcon][OtherCar|SocioEcon]",
   "[VehicleYear|SocioEcon:RiskAversion]",
   "[MakeModel|SocioEcon:RiskAversion]",
   "[SeniorTrain|Age:RiskAversion]",
   "[HomeBase|SocioEcon:RiskAversion]",
   "[AntiTheft|SocioEcon:RiskAversion]",
   "[RuggedAuto|VehicleYear:MakeModel]",
   "[Antilock|VehicleYear:MakeModel]",
   "[DrivingSkill|Age:SeniorTrain]",
   "[CarValue|VehicleYear:MakeModel:Mileage]",
   "[Airbag|VehicleYear:MakeModel]",
   "[DrivQuality|RiskAversion:DrivingSkill]",
   "[Theft|CarValue:HomeBase:AntiTheft]",
   "[Cushioning|RuggedAuto:Airbag]",
   "[DrivHist|RiskAversion:DrivingSkill]",
   "[Accident|DrivQuality:Mileage:Antilock]",
   "[ThisCarDam|RuggedAuto:Accident]",
   "[OtherCarCost|RuggedAuto:Accident]",
   "[MedCost|Age:Accident:Cushioning]",
   "[ILiCost|Accident]",
   "[ThisCarCost|ThisCarDam:Theft:CarValue]",
   "[PropCost|ThisCarCost:OtherCarCost]")
dag = model2network(modelstring)
ins_fit <- bn.fit(dag,insurance)
graphviz.plot(dag)

pie_fitted_plot(ins_fit,"GoodStudent")
}
